Computing Elo rank properly and rounding it to int in the UI. In order to bind to port 80, first set GOPATH for the root user.

+++ foosrank/models.go
@@ -21,7 +21,7 @@ type RankedPlayer struct {
 
 
 type EloRank struct {
-	Value int
+	Value float64
 }
 
-type RankingFunction func(int, int) (int, int)
+type RankingFunction func(float64, float64) (float64, float64)
